intro/backend/requirements.txt
flask
pillow
pytesseract
pdf2image

bcrypt